Zipfile issue
Occasionally, you get the zipfile that when you unzip it, it has all the
images in a sub-directory instead of being right in the root directory 
of the zipfile.

For example, my icons.zip was actually icons/icons/(files) when I unzip it - silly windows!). 

If you're not going to allow recursion of all directories, then at least 
check one directory beneath what you just unzipped to see if there's 
anything there.